So the bug was happening ever since added threading. The correct way to do it is to call async instead of launch and await the new list, then set the contents of the state list in the main UI thread.
I think the reason the isLoading guard wasn't working was because the UI thread doesn't trigger recomposition until maybe the sort is under way already? Also maybe the recomposition starts from the inside-out (or in parallel). Since the column entries don't use isLoading, they're recomposed when the mutableList changes and don't care if the guard is.

The effect was more evident if the guard was passed as a State<Boolean>. The lazy list doesn't realize the guard is unset, and the progress indicators spin forever.

Idk if this is related: https://issuetracker.google.com/issues/163069767?pli=1