Documentation and add back loading guards.
Was dealing with a bug caused by generating the transaction json from the release build of parthenos, but the setprecision fix was only built for debug build. Somehow this only affected Alhena though, and the json still showed several 0s after the decimal point, it was just that the value was rounded to 6 sig figs.